Moved discovered channels endpoint to channels module
